fix repositories
