+++ app/Http/Controllers/API/V1/AirtimeController.php
@@ -9,7 +9,7 @@ use App\Enums\TransactionType;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\AirtimeRequest;
 use App\Models\AirtimeAccount;
-use App\Repositories\V2\TransactionRepository;
+use App\Repositories\TransactionRepository;
 use App\Services\SidoohAccounts;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
