+++ app/Http/Controllers/API/V1/MerchantController.php
@@ -9,7 +9,7 @@ use App\Enums\ProductType;
 use App\Enums\TransactionType;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\MerchantRequest;
-use App\Repositories\V2\TransactionRepository;
+use App\Repositories\TransactionRepository;
 use App\Services\SidoohAccounts;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Log;
