+++ app/Http/Controllers/API/V1/UtilityController.php
@@ -9,7 +9,7 @@ use App\Enums\TransactionType;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\ProductRequest;
 use App\Models\UtilityAccount;
-use App\Repositories\V2\TransactionRepository;
+use App\Repositories\TransactionRepository;
 use App\Services\SidoohAccounts;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
