+++ app/Http/Controllers/API/V1/WithdrawController.php
@@ -10,7 +10,7 @@ use App\Enums\TransactionType;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\EarningRequest;
 use App\Models\EarningAccount;
-use App\Repositories\V2\TransactionRepository;
+use App\Repositories\TransactionRepository;
 use App\Services\SidoohAccounts;
 use Exception;
 use Illuminate\Http\JsonResponse;
