+++ app/Repositories/TransactionRepository.php
@@ -2,15 +2,14 @@
 
 namespace App\Repositories;
 
+use App\DTOs\PaymentDTO;
 use App\Enums\Description;
 use App\Enums\EarningAccountType;
 use App\Enums\EventType;
-use App\Enums\MerchantType;
 use App\Enums\PaymentMethod;
 use App\Enums\PaymentSubtype;
 use App\Enums\ProductType;
 use App\Enums\Status;
-use App\Enums\TransactionType;
 use App\Helpers\Product\Purchase;
 use App\Helpers\Tanda\TandaApi;
 use App\Models\EarningAccount;
@@ -22,10 +21,10 @@ use App\Services\SidoohNotify;
 use App\Services\SidoohPayments;
 use App\Services\SidoohSavings;
 use App\Traits\ApiResponse;
+use Error;
 use Exception;
 use Illuminate\Auth\AuthenticationException;
-use Illuminate\Support\Arr;
-use Illuminate\Support\Collection;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use Throwable;
 
@@ -39,198 +38,211 @@ class TransactionRepository
      * @throws AuthenticationException
      * @throws Throwable
      */
-    public static function createTransactions(array $transactionsData, $data): array
+    public static function createTransaction(array $transactionData, $data): Transaction
     {
-        $transactions = collect();
+        $transaction = Transaction::create($transactionData);
 
-        foreach ($transactionsData as $transactionData) {
-            $transactions->add(Transaction::create($transactionData));
-        }
+        self::initiatePayment($transaction, $transactionData['account'], $data);
 
-        self::initiatePayment($transactions, $data);
-
-        return Arr::pluck($transactions, 'id');
+        return $transaction;
     }
 
     /**
      * @throws AuthenticationException
-     * @throws \Throwable
+     * @throws Throwable
      */
-    public static function initiatePayment(Collection $transactions, array $data): void
+    public static function initiatePayment(Transaction $t, array $account, array $data): void
     {
-        if (isset($data['debit_account'])) {
-            $debit_account = $data['debit_account'];
-        } else {
-            $account = $data['payment_account'];
-            $debit_account = $data['method'] === PaymentMethod::MPESA ? $account['phone'] : $account['id'];
+        $paymentMethod = $data['method'];
+
+        $debit_account = $data['debit_account'] ?? match ($paymentMethod) {
+            PaymentMethod::MPESA => $account['phone'],
+            PaymentMethod::VOUCHER => SidoohPayments::findSidoohVoucherIdForAccount($account['id'])
+        };
+
+        $paymentData = new PaymentDTO($t->account_id, $t->amount, $t->description, $t->destination, $paymentMethod, $debit_account);
+
+        if (is_int($t->product_id)) {
+            $t->product_id = ProductType::tryFrom($t->product_id);
         }
 
-        $transactionsData = $transactions->map(fn($t) => [
-            'reference'   => $t->id,
-            'product_id'  => $t->product_id,
-            'amount'      => $t->amount,
-            'destination' => $t->destination,
-            'description' => $t->description,
-        ]);
+        match ($t->product_id) {
+            ProductType::VOUCHER => $paymentData->setVoucher(SidoohPayments::findSidoohVoucherIdForAccount(SidoohAccounts::findByPhone($t->destination)['id'])),
+            ProductType::MERCHANT => $paymentData->setMerchant($data['merchant_type'], $data['business_number'], $data['account_number'] ?? ''),
+            default => $paymentData->setDestination(PaymentMethod::FLOAT, 1)
+        };
 
-        $responseData = SidoohPayments::requestPayment($transactionsData, $data['method'], $debit_account);
+        $p = SidoohPayments::requestPayment($paymentData);
 
-        // TODO: Revert this to: if (!isset($response["data"]["payments"])) throw new Exception("Purchase Failed!");
-        //  Reason may not be due to payment failure, could be a connection issue etc...
-        //  We would then have to manually check. Or implement a query endpoint that polls payment srv at set intervals
-        if (! isset($responseData['payments'])) {
-//            $transactions->each(fn($t) => $t->update(['status' => Status::FAILED]));
+        $paymentData = [
+            'transaction_id' => $t->id,
+            'payment_id'     => $p['id'],
+            'amount'         => $p['amount'],
+            'type'           => $p['type'],
+            'subtype'        => $p['subtype'],
+            'status'         => $p['status'],
+            'extra'          => [
+                'debit_account' => $debit_account,
+                ...($p['destination'] ?? []),
+            ],
+        ];
 
-            throw new Exception('Purchase Failed!');
-        }
+        Payment::create($paymentData);
 
-        $paymentData = array_map(function($p) use ($responseData, $debit_account) {
-            return [
-                'transaction_id' => $p['reference'],
-                'payment_id'     => $p['id'],
-                'amount'         => $p['amount'],
-                'type'           => $p['type'],
-                'subtype'        => $p['subtype'],
-                'status'         => $p['status'],
-                'extra'          => json_encode($responseData['debit_voucher'] ?? ['debit_account' => $debit_account]),
-                'created_at'     => now(),
-                'updated_at'     => now(),
-            ];
-        }, $responseData['payments']);
-        Payment::insert($paymentData);
-
-        if ($responseData && $data['method'] === PaymentMethod::VOUCHER) {
-            self::requestPurchase($transactions, $responseData);
+        if ($p && $data['method'] === PaymentMethod::VOUCHER && $t->product_id !== ProductType::MERCHANT) {
+            self::requestPurchase($t);
         }
     }
 
     /**
      * @throws Throwable
      */
-    public static function requestPurchase(Collection $transactions, array $paymentsData): void
+    public static function requestPurchase(Transaction $transaction): void
     {
         try {
-            foreach ($transactions as $transaction) {
-                $purchase = new Purchase($transaction);
+            $purchase = new Purchase($transaction);
 
-                if (is_int($transaction->product_id)) {
-                    $transaction->product_id = ProductType::tryFrom($transaction->product_id);
-                }
-
-                match ($transaction->product_id) {
-                    ProductType::AIRTIME      => $purchase->airtime(),
-                    ProductType::UTILITY      => $purchase->utility(),
-                    ProductType::SUBSCRIPTION => $purchase->subscription(),
-                    ProductType::VOUCHER      => $purchase->voucher($paymentsData),
-                    default                   => throw new Exception('Invalid product purchase!'),
-                };
+            if (is_int($transaction->product_id)) {
+                $transaction->product_id = ProductType::tryFrom($transaction->product_id);
             }
+
+            match ($transaction->product_id) {
+                ProductType::AIRTIME => $purchase->airtime(),
+                ProductType::UTILITY => $purchase->utility(),
+                ProductType::SUBSCRIPTION => $purchase->subscription(),
+                ProductType::VOUCHER => $purchase->voucher(),
+                ProductType::MERCHANT => $purchase->merchant(),
+                default => throw new Exception('Invalid product purchase!'),
+            };
         } catch (Exception $err) {
             Log::error($err);
         }
     }
 
+    public static function handleFailedPayment(Transaction $transaction, Request $payment): void
+    {
+        $transaction->payment->update(['status' => Status::FAILED]);
+        $transaction->status = Status::FAILED;
+        $transaction->save();
+
+        if ($payment->subtype === PaymentSubtype::STK->name && isset($payment->code)) {
+            $message = match ($payment->code) {
+                1 => 'You have insufficient Mpesa Balance for this transaction. Kindly top up your Mpesa and try again.',
+                default => 'Sorry! We failed to complete your transaction. No amount was deducted from your account. We apologize for the inconvenience. Please try again.',
+            };
+        } elseif (ProductType::tryFrom($transaction->product_id) === ProductType::MERCHANT) {
+            $destination = $transaction->destination;
+            $message = "Sorry! We failed to complete your transaction to merchant: $destination. No amount was deducted from your account. We apologize for the inconvenience. Please try again.";
+        } else {
+            $message = 'Sorry! We failed to complete your transaction. No amount was deducted from your account. We apologize for the inconvenience. Please try again.';
+        }
+
+        $account = SidoohAccounts::find($transaction->account_id);
+
+        SidoohNotify::notify([$account['phone']], $message, EventType::PAYMENT_FAILURE);
+    }
+
+    /**
+     * @throws Throwable
+     */
+    public static function handleCompletedPayment(Transaction $transaction): void
+    {
+        $transaction->payment->update(['status' => Status::COMPLETED]);
+
+        TransactionRepository::requestPurchase($transaction);
+    }
+
+
     /**
      * @throws AuthenticationException
      * @throws Throwable
      */
-    public static function createWithdrawalTransactions(array $transactionsData, $data): array
+    public static function createWithdrawalTransaction(array $transactionData, $data): Transaction
     {
-        $transactions = collect();
-        foreach ($transactionsData as $transactionData) {
-            $transactions->add(Transaction::create($transactionData));
-        }
+        $transaction = Transaction::create($transactionData);
 
-        self::initiateSavingsWithdrawal($transactions, $data);
+        self::initiateSavingsWithdrawal($transaction, $data);
 
-        return Arr::pluck($transactions, 'id');
+        return $transaction;
     }
 
     /**
      * @throws \Throwable
      */
-    public static function initiateSavingsWithdrawal(Collection $transactions, array $data): Collection
+    public static function initiateSavingsWithdrawal(Transaction $transaction, array $data): void
     {
-        $responses = SidoohSavings::withdrawEarnings($transactions, $data['method']);
-
-        $transactions->each(function($tx) use ($responses) {
-            if (array_key_exists($tx->id, $responses['failed'])) {
-                $response = $responses['failed'][$tx->id];
-
-                SavingsTransaction::create([
-                    'transaction_id' => $tx->id,
-                    'description'    => $response,
-                    'type'           => TransactionType::DEBIT,
-                    'amount'         => $tx->amount,
-                    'status'         => Status::FAILED,
-                ]);
-
-                $tx->status = Status::FAILED;
-                $tx->save();
-            }
-
-            if (array_key_exists($tx->id, $responses['completed'])) {
-                $response = $responses['completed'][$tx->id];
-
-                SavingsTransaction::create([
-                    ...$response,
-                    'reference'      => $response['id'],
-                    'transaction_id' => $tx->id,
-                    'id'             => null,
-                ]);
-
-                //TODO: Fix for new users.
-                $acc = EarningAccount::firstOrCreate([
-                    'type'       => EarningAccountType::WITHDRAWALS->name,
-                    'account_id' => $tx->account_id,
-                ]);
-                $acc->update(['self_amount' => $acc->self_amount + $tx->amount]);
-
-                $tx->refresh();
-            }
-        });
-
-        return $transactions;
-    }
-
-    public static function handleFailedPayments(Collection $transactions, Collection $failedPayments): void
-    {
-        $transactions->each(function($transaction) use ($failedPayments) {
-            $transaction->payment->update(['status' => Status::FAILED]);
+        try {
+            $response = SidoohSavings::withdrawEarnings($transaction, $data['method']);
+        } catch (Exception) {
             $transaction->status = Status::FAILED;
             $transaction->save();
 
-            $result = $failedPayments->firstWhere('id', $transaction->payment->payment_id);
+            throw new Error('Something went wrong, please try again later.');
+        }
 
-            if ($result['subtype'] === PaymentSubtype::STK->name && isset($result['stk_result_code'])) {
-                $message = match ($result['stk_result_code']) {
-                    1       => 'You have insufficient Mpesa Balance for this transaction. Kindly top up your Mpesa and try again.',
-                    default => 'Sorry! We failed to complete your transaction. No amount was deducted from your account. We apologize for the inconvenience. Please try again.',
-                };
-            } elseif (ProductType::tryFrom($transaction->product_id) === ProductType::MERCHANT) {
-                $destination = $transaction->destination;
-                $message = "Sorry! We failed to complete your transaction to merchant: $destination. No amount was deducted from your account. We apologize for the inconvenience. Please try again.";
-            } else {
-                $message = 'Sorry! We failed to complete your transaction. No amount was deducted from your account. We apologize for the inconvenience. Please try again.';
-            }
+        SavingsTransaction::create([
+            'transaction_id' => $transaction->id,
+            'savings_id'     => $response['id'],
+            'amount'         => $response['amount'],
+            'description'    => $response['description'],
+            'type'           => $response['type'],
+            'status'         => $response['status'],
+            'extra'          => $response['extra'],
+        ]);
 
-            $account = SidoohAccounts::find($transaction->account_id);
+        //TODO: Fix for new users.
+        $acc = EarningAccount::firstOrCreate([
+            'type'       => EarningAccountType::WITHDRAWALS->name,
+            'account_id' => $transaction->account_id,
+        ]);
+        $acc->increment('self_amount', $transaction->amount);
 
-            SidoohNotify::notify([$account['phone']], $message, EventType::PAYMENT_FAILURE);
-        });
+
+        $tagline = config('services.sidooh.tagline');
+        $message = "Your withdrawal request has been received. Please be patient as we review it.\n\n$tagline";
+
+        $account = SidoohAccounts::find($transaction->account_id);
+
+        SidoohNotify::notify([$account['phone']], $message, EventType::WITHDRAWAL_PAYMENT);
+    }
+
+    public static function handleFailedWithdrawal(Transaction $transaction, Request $savings): void
+    {
+        $transaction->savingsTransaction->update(['status' => Status::FAILED]);
+
+        EarningAccount::accountId($transaction->account_id)
+            ->withdrawal()
+            ->first()
+            ->decrement('self_amount', $transaction->amount);
+
+        $transaction->status = Status::FAILED;
+        $transaction->save();
+
+        $message = 'Sorry! We failed to complete your withdrawal request. No amount was deducted from your account. We apologize for the inconvenience. Please try again.';
+
+        $account = SidoohAccounts::find($transaction->account_id);
+
+        SidoohNotify::notify([$account['phone']], $message, EventType::WITHDRAWAL_FAILURE);
     }
 
     /**
-     * @throws \Throwable
+     * @throws Throwable
      */
-    public static function handleCompletedPayments(Collection $transactions, Collection $completedPayments, array $requestData = []): void
+    public static function handleCompletedWithdrawal(Transaction $transaction): void
     {
-        $ids = $completedPayments->pluck('id');
+        $transaction->savingsTransaction->update(['status' => Status::COMPLETED]);
 
-        $payments = Payment::whereIn('payment_id', $ids);
-        $payments->update(['status' => Status::COMPLETED]);
+        $transaction->status = Status::COMPLETED;
+        $transaction->save();
 
-        TransactionRepository::requestPurchase($transactions, $requestData);
+        $destination = $transaction->savingsTransaction->extra['destination'];
+        $account = $transaction->savingsTransaction->extra['destination_account'];
+        $message = "Congrats! Your have withdrawn $transaction->amount points from your earnings account to $destination - $account successfully.\n";
+        $message .= config('services.sidooh.tagline');
+
+        $account = SidoohAccounts::find($transaction->account_id);
+
+        SidoohNotify::notify([$account['phone']], $message, EventType::WITHDRAWAL_PAYMENT);
     }
 
     public static function checkRequestStatus(Transaction $transaction, string $requestId): void
@@ -242,9 +254,6 @@ class TransactionRepository
         };
     }
 
-    /**
-     * @throws \Illuminate\Auth\AuthenticationException|Exception
-     */
     public static function refundTransaction(Transaction $transaction): void
     {
         $phone = SidoohAccounts::find($transaction->account_id)['phone'];
@@ -273,75 +282,5 @@ class TransactionRepository
         };
 
         SidoohNotify::notify([$phone], $message, EventType::VOUCHER_REFUND);
-    }
-
-    /**
-     * @throws AuthenticationException
-     * @throws Throwable
-     */
-    public static function createB2bTransaction(array $transactionData, array $data): int
-    {
-        $transaction = Transaction::create($transactionData);
-
-        self::initiateB2bPayment($transaction, $data);
-
-        return $transaction->id;
-    }
-
-    /**
-     * @throws AuthenticationException
-     * @throws \Throwable
-     */
-    public static function initiateB2bPayment(Transaction $transaction, array $data): void
-    {
-        if (isset($data['debit_account'])) {
-            $debit_account = $data['debit_account'];
-        } else {
-            $account = $data['payment_account'];
-            $debit_account = $data['method'] === PaymentMethod::MPESA ? $account['phone'] : $account['id'];
-        }
-
-        $transactionData = [
-            'reference'   => $transaction->id,
-            'product_id'  => $transaction->product_id,
-            'amount'      => $transaction->amount,
-            'destination' => $transaction->destination,
-            'description' => $transaction->description,
-        ];
-
-        $merchantDetails = [
-            'merchant_type' => $data['merchant_type'],
-        ];
-
-        if ($data['merchant_type'] === MerchantType::MPESA_PAY_BILL) {
-            $merchantDetails += [
-                'paybill_number' => $data['business_number'],
-                'account_number' => $data['account_number'],
-            ];
-        } else {
-            $merchantDetails += [
-                'till_number'    => $data['business_number'],
-                'account_number' => '',
-            ];
-        }
-
-        $responseData = SidoohPayments::requestB2bPayment($transactionData, $data['method'], $debit_account, $merchantDetails);
-
-        if (! isset($responseData['payments']) && ! isset($responseData['b2b_payment'])) {
-            throw new Exception('Purchase Failed!');
-        }
-
-        $p = $responseData['b2b_payment'];
-        Payment::insert([
-            'transaction_id' => $p['reference'],
-            'payment_id'     => $p['id'],
-            'amount'         => $p['amount'],
-            'type'           => $p['type'],
-            'subtype'        => $p['subtype'],
-            'status'         => $p['status'],
-            'extra'          => json_encode($responseData['debit_voucher'] ?? ['debit_account' => $debit_account]),
-            'created_at'     => now(),
-            'updated_at'     => now(),
-        ]);
     }
 }
